requirements.txt
allennlp==2.3.0
allennlp-models==2.3.0
xmltodict
portion
tensorflow==2.9.0
keras==2.9.0
numpy==1.21.5
transformers==4.26.0
spacy==2.3.0
sentence-transformers
streamlit-chat==0.0.2.1
streamlit==1.22.0
streamlit-searchbox==0.1.3
pdfminer.six==20221105
beautifulsoup4==4.12.2
torch==1.9.0+cu111